html/semantics/forms/the-label-element/label-attrubutes.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTML Test: The label element</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<form id="fm" style="display:none">
  <label id="lbl0" for="test0"></label>
  <b id="test0"></b>

  <label id="lbl1" for="test1"></label>
  <input id="test1" class="test1"></input>
  <label id="lbl2" for="test1"></label>

  <label id="lbl3">
    <a></a>
    <input id="input1" class="textbox">
    <input id="input2">
  </label>
</form>

<script>

  test(function () {
    assert_not_equals(document.getElementById("lbl0").control, document.getElementById("test0"),
                      "The b element is not a labelable element");
    assert_equals(document.getElementById("lbl0").control, null,
                  "There is no element is labeled by lbl0.");
  }, "unlabelable element matches the for attribute");

  test(function () {
    var labels = document.querySelector(".test1").labels;
    assert_true(labels instanceof NodeList, "labels should be an instance of NodeList.");
    assert_equals(labels.length, 2, "The control should have three labels.");
    assert_array_equals(labels, [document.getElementById("lbl1"), document.getElementById("lbl2")],
                        "The labels should be sorted in tree order.");
  }, "mutiple labels for one control");
  

  test(function () {
    assert_equals(document.getElementById("lbl3").control, document.querySelector(".textbox"),
                  "The labeled control should be the first labelable descendant element.");
  }, "label for the first labelable descendant element");
  
  test(function () {
    assert_equals(document.getElementById("lbl0").form, document.getElementById("fm"),
                  "the label.form should return its form owner.'");
  }, "label's form owner");

</script>
